refactor(cleanup): update devspace controller module name
