+++ pkg/controllers/devspace/attach.go
@@ -1,4 +1,4 @@
-package robot
+package devspace
 
 import (
 	"context"
