+++ pkg/controllers/devspace/check.go
@@ -1,4 +1,4 @@
-package robot
+package devspace
 
 import (
 	"context"
